tools: add capability to provide make command via environment
Not all operating systems name the GNU Make `make`. FreeBSD e.g. uses a
different dialect of Make, that seems to be incompatible with GNU make.
(I wasn't able to get `make` run, but `gmake` works).

This allows our test scripts to be configured via the environment
variable `MAKE` to point to a different make command.